Patch by Børre Gaup to clean corrupted xml data returned from Libarchive when reading xml data from a zhfst file stored in RAM:
libarchive seems to add one char from the following file in the stream pr whitespace char it (erroneously) removes from the xml data around newlines, since the returned xml file is truncated compared to the original file length reported when stored in the archive. Although the xml info set is exactly the same, the extra chars after the final closing element causes TinyXML2 to crash.

This change removes everything after the final '>', which should clean the string representation of the xml document properly in all cases but one: when the semi random noise contains an '>'. This situation is presently not handled.

git-svn-id: svn://svn.code.sf.net/p/hfst/code/trunk@3653 941e2c2b-deac-454f-805a-451daa25f33c